Change link of expired session to today's rather than tomorrow's session
Currently, when you have done the dishes, you have to go to the (then already closed) session via the sessions button in the header, which adds an unnecessary action to the procedure. I would imagine very few people use the front-page widget to go to the session of tomorrow, so it might be changed to link to today's session, to make things easier for the ones who have done the dishes.